api/mraa/spi.h Merged in latest master and applied same changes to spi_write_word so that resultcode of -1 works api/mraa/spi.hpp Merged in latest master and applied same changes to spi_write_word so that resultcode of -1 works src/spi/spi.c Implemented soft_cs, allows devices with 8/9bit drivers to do 16bit spi transfers. Merged in latest master and applied same changes to spi_write_word

+++ api/mraa/spi.h
@@ -143,7 +143,7 @@ int mraa_spi_write(mraa_spi_context dev, uint8_t data);
  * @param data Data to send
  * @return Data received on the miso line
  */
-uint16_t mraa_spi_write_word(mraa_spi_context dev, uint16_t data);
+long mraa_spi_write_word(mraa_spi_context dev, uint16_t data);
 
 /**
  * Write Buffer of bytes to the SPI device. The pointer return has to be
